Copy Ruby files into repo
Ttycaca is an outgrowth of my work on both usaidwat and downterm, which
both require the ability to query the terminal for information. Both
usaidwat and downterm use highline for that, but occasionally I don't
keep the versions of highline they use in sync, which causes problems;
having a common way to query the terminal for information will be
incredibly useful.